Remove checking of filesystem type from hasquota()
The checking of filesystem type in verification whether a filesystem
has quotas has been sporadically causing us problems with new filesystems
or so. The checking isn't actually very useful because in (rather unlikely)
worst case if some filesystem has valid quota options but uses quota system
which quota-tools do not understand, we try to work with quota files and
the header check fails. So it's not a big deal.

Signed-off-by: Jan Kara <jack@suse.cz>