quota_nld: Fix usage of netlink multicast groups
quota_nld has used VFS_DQUOT netlink family id as an id of a multicast
group to listen to. That is wrong - instead proper multicast group id
should be looked up. To do that we need a support for this functionality
in libnl so convert quota_nld to use libnl-3.0 and properly lookup
multicast group id.

Signed-off-by: Jan Kara <jack@suse.cz>