Don't link all binaries with ldap library
The default action-if-found of AC_CHECK_LIB() is to append checked
library to LIBS. Thus check for ldap library resulted in unwanted
addition of -lldap to LIBS as [] is an empty string in M4 and the
default action is used.

Fix the problem by providing proper action-if-found which was currently
just hidden behind the check. Also do similar cleanup for
AC_CHECK_HEADER check although there it didn't have any undesired
side-effect.

Reported-by: Petr Písař <petrp@users.sf.net>
Signed-off-by: Jan Kara <jack@suse.cz>