Make messages about failures for NFS consistent with local filesystems
Currently, some types for failures when fetching quota information for
NFS filesystem were silent (e.g. when rpc.rquotad was not running) while
others were reporting error message (e.g. when rpc connection failed).
There's no big difference in these for the user / administrator and also
is inconsistent with how we deal with local filesystems - there we
report error if the filesystem was explicitely specified on command line
and silently ignore it for "scan all" operations.

So change error reporting for NFS to report errors about quota not being
supported if and only if filesystem was explicitely specified on command
line.

Signed-off-by: Jan Kara <jack@suse.cz>