quotacheck: Skip checking of filesystems with hidded quota files early
Currently when filesystem tracks quotas in hidden quota files (but is
also capable of tracking it in regular files), we scan the whole
filesystem and only after that find out we cannot actually make use of
newly created quota files. Just detect this case before scanning the fs
and skip it.

Signed-off-by: Jan Kara <jack@suse.cz>