edquota: Fix editing of individual user grace times
When user is not over softlimit, there's no point to set the grace time.
The kernel will just ignore it. Output error message instead so that
user is not confused. Also editing of grace times when quotas are
enabled did not really work because we didn't tell the kernel grace
times have changed. Fix it.

Signed-off-by: Jan Kara <jack@suse.cz>