Provide quotactl_mnt() helper and use it where possible
There are lot of places which call quotactl(2) on a particular mount
point. Provide quotactl_mnt() helper instead of having to call
do_quotactl() and having to figure out whether devname is usable or not
based on fs type. Also use quotactl_handle() in a few places which were
still using do_quotactl().

Signed-off-by: Jan Kara <jack@suse.cz>